requirements.txt
pymongo==3.6.0
cassandra-driver==3.12.0
progressbar2==3.36.0
psutil==5.4.5
tqdm==4.22.0
cachetools==2.0.1
